Move pkg/imagevector to imagevector
